requirements.txt
flask==2.2.5
gunicorn==20.1.0
python-dotenv==0.21.1